pyproject.toml
[project]
name = "xunlong"
version = "0.1.0"
description = "AI-powered PowerPoint generation system"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.12.0",
    "openai>=1.0.0",
    "jinja2>=3.1.0",
    "loguru>=0.7.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "pytz>=2023.0",
    "playwright>=1.57.0",
    "aiofiles>=25.1.0",
    "pillow>=12.1.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "instructor>=1.13.0",
    # FastAPI相关依赖
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]
